requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
PyPDF2==3.0.1
PyMuPDFb==1.23.6
pytesseract==0.3.10
python-magic==0.4.27; 'linux' in sys.platform
python-magic-bin==0.4.14; sys_platform == 'win32'
python-pptx==0.6.23
XlsxWriter==3.1.9